Add management of pinentry error
In the case there is an error in the first call to pass (entries listing)
the procedure is interrupted.

Given that most users will have a gpgagent active managing the first
call should be enough.

However to be more stable maybe pass error management should be included at
each call.

Solves Issue #1.